% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMGIG.R
\name{fMGIG}
\alias{fMGIG}
\title{Return the unnormalized density of MGIG}
\usage{
fMGIG(X, nu = 6, phi = diag(3), psi = diag(3), logirithm = TRUE)
}
\arguments{
\item{X}{a positive definite matrix where to calculate density}

\item{nu}{double, nu parameter for the MGIG}

\item{phi}{phi parameter positive (semi)defined matrix, reduce to inverse-Wishart if set to 0}

\item{psi}{psi parameter positive (semi)defined matrix, reduce to Wishart if set to 0}

\item{logirithm}{whether to take log}
}
\value{
a matrix, the mode of the MGIG
}
\description{
The unnormalized density of MGIG given X
}
\examples{
my.mode <- fMGIG(X = diag(3), nu = 6, phi = diag(3), psi = diag(3))
}
