% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMGIG.R
\name{KLdiv}
\alias{KLdiv}
\title{Estimate the KL divergence between two MGIG distribution}
\usage{
KLdiv(nu1, phi1, psi1, nu2, phi2, psi2, n_samples = 5000)
}
\arguments{
\item{nu1}{double, nu parameter for the first MGIG}

\item{phi1}{phi parameter for the first positive (semi)defined matrix, reduce to inverse-Wishart if set to 0}

\item{psi1}{psi parameter for the first positive (semi)defined matrix, reduce to Wishart if set to 0}

\item{nu2}{double, nu parameter for the second MGIG}

\item{phi2}{phi parameter for the second positive (semi)defined matrix, reduce to inverse-Wishart if set to 0}

\item{psi2}{psi parameter for the second positive (semi)defined matrix, reduce to Wishart if set to 0}

\item{n_samples}{samples to take to estimate the KL divergence}
}
\value{
a matrix, the mode of the MGIG
}
\description{
Calculate the KL(MGIG(nu1,phi1,psi1)||MGIG(nu2,phi2,psi2))
}
