% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMGIG.R
\name{mMGIG}
\alias{mMGIG}
\title{Return the mode of Matrix Generalized Inverse Gaussian distribution}
\usage{
mMGIG(nu = 6, phi = diag(3), psi = diag(3))
}
\arguments{
\item{nu}{double, nu parameter for the MGIG}

\item{phi}{phi parameter positive (semi)defined matrix, reduce to inverse-Wishart if set to 0}

\item{psi}{psi parameter positive (semi)defined matrix, reduce to Wishart if set to 0}
}
\value{
a matrix, the mode of the MGIG
}
\description{
Mode of Matrix Generalized Inverse Gaussian distribution, 
Method follow Farideh Fazayeli and Arindam Banerjee 2016
"The Matrix Generalized Inverse Gaussian Distribution:
Properties and Applications" 
by solving the ARE
}
\examples{
my.mode <- mMGIG(nu = 6, phi = diag(3), psi = diag(3))
}
