% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMGIG.R
\name{rMGIG}
\alias{rMGIG}
\title{Sampling from Matrix Generalized Inverse Gaussian distribution}
\usage{
rMGIG(
  n = 1,
  nu = 6,
  phi = diag(3),
  psi = diag(3),
  df = 6,
  list = TRUE,
  maxit = 10000
)
}
\arguments{
\item{n}{number of sample needed}

\item{nu}{double, nu parameter for the MGIG}

\item{phi}{phi parameter positive (semi)defined matrix, reduce to inverse-Wishart if set to 0}

\item{psi}{psi parameter positive (semi)defined matrix, reduce to Wishart if set to 0}

\item{df}{positive double, should be greater than dimision-1 degree of freedom for the proposal Wishart distribution during importance sampling}

\item{list}{bool, whether want list output, if ture will return list of full matrix otherwise return matrix of upper triagular part}

\item{maxit}{integer maximum iterations for importance sampling algorithm}
}
\value{
a matrix with upper triangular part of the positive defined matrix as columns or a list whose entries are the sampled matrix
}
\description{
Sampling from Matrix Generalized Inverse Gaussian distribution, 
Method follow Farideh Fazayeli and Arindam Banerjee 2016
"The Matrix Generalized Inverse Gaussian Distribution:
Properties and Applications"
}
\examples{
my.sample <- rMGIG(n = 1, nu = 6, phi = diag(3), 
psi = diag(3), df = 6, 
list = FALSE, maxit = 1000)
my.sample.list <- rMGIG(n = 1, nu = 6, phi = diag(3), 
psi = diag(3), df = 6, 
list = TRUE, maxit = 1000)
}
