\name{rMGIG-package}
\alias{rMGIG-package}
\alias{rMGIG}
\docType{package}
\title{
\packageTitle{rMGIG}
}
\description{
\packageDescription{rMGIG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rMGIG}
\packageIndices{rMGIG}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rMGIG}

Maintainer: \packageMaintainer{rMGIG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
